// frontend/generated/models/TransactionStatusDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TransactionStatusDto = {
    Pending: 'PENDING',
    Success: 'SUCCESS',
    Failed: 'FAILED'
} as const;
export type TransactionStatusDto = typeof TransactionStatusDto[keyof typeof TransactionStatusDto];


export function instanceOfTransactionStatusDto(value: any): boolean {
    for (const key in TransactionStatusDto) {
        if (Object.prototype.hasOwnProperty.call(TransactionStatusDto, key)) {
            if (TransactionStatusDto[key as keyof typeof TransactionStatusDto] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TransactionStatusDtoFromJSON(json: any): TransactionStatusDto {
    return TransactionStatusDtoFromJSONTyped(json, false);
}

export function TransactionStatusDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TransactionStatusDto {
    return json as TransactionStatusDto;
}

export function TransactionStatusDtoToJSON(value?: TransactionStatusDto | null): any {
    return value as any;
}

export function TransactionStatusDtoToJSONTyped(value: any, ignoreDiscriminator: boolean): TransactionStatusDto {
    return value as TransactionStatusDto;
}